+++ app-modules/integration-ai/src/Client/Concerns/InitializesClient.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Assist\IntegrationAI\Client\Concerns;
+
+trait InitializesClient
+{
+    abstract protected function initializeClient(): void;
+}
